Move HostListItem from host.go to command.go
